Fix: functions update and delete of all controllers and lint of all files

+++ backend/src/app/controllers/EstablishmentController.js
@@ -14,23 +14,23 @@ class EstablishmentController {
         cnpj,
         address,
         addressComplement = "",
-        phones 
+        phones
       } = req.body;
-  
+
       const emailExist = await Establishment.findOne({ email });
-  
+
       if (emailExist) {
         return res.status(400).json({ message: 'E-mail já existente.' });
       }
-  
+
       const _id = await crypto.randomBytes(4).toString('HEX');
-  
+
       const token = await jwt.sign({ _id }, authConfig.secret, {
         expiresIn: authConfig.expiresIn,
       });
-  
+
       const passwordHash = await bcrypt.hash(password, 8);
-  
+
       const establishment = await Establishment.create({
         _id,
         name,
@@ -43,13 +43,13 @@ class EstablishmentController {
         last_login: new Date(),
         token,
       });
-  
+
       const createdData = new Date(establishment.createdAt.valueOf() - establishment.createdAt.getTimezoneOffset() * 60000);
       const data_criacao = createdData.toISOString().replace(/\.\d{3}Z$/, '');
-  
+
       const updatedData = new Date(establishment.updatedAt.valueOf() - establishment.updatedAt.getTimezoneOffset() * 60000);
       const data_atualizacao = updatedData.toISOString().replace(/\.\d{3}Z$/, '');
-  
+
       return res.status(201).json({
         _id: establishment.id,
         nome: establishment.name,
@@ -64,8 +64,8 @@ class EstablishmentController {
         data_criacao,
         data_atualizacao,
         token: establishment.token,
-      });  
-    } catch(err) {
+      });
+    } catch (err) {
       return res.status(400).send(err);
     }
   }
@@ -76,13 +76,13 @@ class EstablishmentController {
 
       const createdData = new Date(establishment.createdAt.valueOf() - establishment.createdAt.getTimezoneOffset() * 60000);
       const data_criacao = createdData.toISOString().replace(/\.\d{3}Z$/, '');
-  
+
       const updatedData = new Date(establishment.updatedAt.valueOf() - establishment.updatedAt.getTimezoneOffset() * 60000);
       const data_ultima_atualizacao = updatedData.toISOString().replace(/\.\d{3}Z$/, '');
-  
+
       const sessionData = new Date(establishment.last_login.valueOf() - establishment.last_login.getTimezoneOffset() * 60000);
       const ultimo_login = sessionData.toISOString().replace(/\.\d{3}Z$/, '');
-  
+
       return res.status(200).json({
         nome: establishment.name,
         email: establishment.email,
@@ -93,31 +93,43 @@ class EstablishmentController {
         data_criacao,
         data_ultima_atualizacao,
         token: establishment.token,
-      });  
-    } catch(err) {
+      });
+    } catch (err) {
       return res.status(400).send(err);
     }
   }
 
   async update(req, res) {
     try {
+      const establishment = await Establishment.findOne({ _id: req.establishmentId });
+
+      if (!establishment) {
+        return res.status(400).json({ message: "Restaurante não encontrado." });
+      }
+
       await Establishment.findOneAndUpdate(
         { _id: req.establishmentId },
         { $set: req.body }
-        );
+      );
 
-      return res.status(200).json({ message: "Dados do restaurante atualizados com sucesso." });  
-    } catch(err) {
+      return res.status(200).json({ message: "Dados do restaurante atualizados com sucesso." });
+    } catch (err) {
       return res.status(400).send(err);
     }
   }
 
   async delete(req, res) {
     try {
+      const establishment = await Establishment.findOne({ _id: req.establishmentId });
+
+      if (!establishment) {
+        return res.status(400).json({ message: "Restaurante não encontrado." });
+      }
+
       await Establishment.findOneAndDelete({ _id: req.establishmentId });
 
       return res.status(200).json({ message: "Restaurante deletado com sucesso." });
-    } catch(err) {
+    } catch (err) {
       return res.status(400).send(err);
     }
   }
